split option
If the xref annotations have multiples in a single line, such as x,y,z, then using the -s X flag will split these on the X value so that instead of 'x,y,z' being the annotation, there will be one for 'x', one for 'y' and one for 'z'. These will then be subject to the -d or -m flags. Also, no need to wrap the separator in quotes. So for a comma, do

,

instead of

','